#352 Implemented comments in PullRequest. added setVersion method with String to not break the api and marked it as deprecated. Renamed Function to getAssetNamesAsBytes to align with other APIs. Added a Test for UnicodeString deserialization. Will move toUTF8Json function to another pull Request.

+++ cip/cip25/src/test/java/com/bloxbean/cardano/client/cip/cip25/NFTMetadataTest.java
@@ -85,7 +85,7 @@ class NFTMetadataTest {
         assertThat(rNft3).isNotNull();
         assertThat(rNftNotAvailable).isNull();
 
-        assertThat(nftMetadata.getVersion()).isEqualTo(BigInteger.valueOf(1));
+        assertThat(nftMetadata.getVersion()).isEqualTo(1);
 
         assertThat(rNft1.getAssetName()).isEqualTo(nft1.getAssetName());
         assertThat(rNft1.getName()).isEqualTo(nft1.getName());
